Move quantity.py and quantity_helper.py to a new quantity module.
